refactor(web): extract components from positions, notifications, news-feed, live-chart pages
