+++ apps/web/src/components/notifications/index.ts
@@ -1,0 +1,15 @@
+export {
+  getNotificationRoute,
+  NotifIcon,
+  iconBg,
+  translateMessage,
+  timeAgo,
+  routeLabel,
+  NotifRow,
+  getMessageKey,
+  NOTIF_TABS,
+  TRADE_KEYS,
+  AGENT_KEYS,
+  PAGE_SIZE,
+} from './components';
+export type { NotifTab } from './components';
