+++ apps/web/src/components/positions/index.ts
@@ -1,0 +1,6 @@
+export {
+  PositionDetailModal,
+  CloseConfirmDialog,
+  LivePnlCell,
+  PositionsTable,
+} from './components';
